feat: implement useSearch hook
